:art: Add diagram of Pangeo Machine Learning Ecosystem in 2023 (#5)
* :art: Add diagram of Pangeo Machine Learning Ecosystem in 2023

Initial draft diagram of the Pangeo ML Ecosystem in 2023. Inspired by Jake VanderPlas's SciPy 2015 Keynote. Using Excalidraw to arrange the logos in-place. Have linked to the sources for all the logos in a README.md file. Also gitignoring png and svg files.

* :mag: Add Pangeo ML Ecosystem 2023 diagram to main README.md

Teaser image for the repository! Also fixing some capitalization and typos in the figures/README.md file.