:art: Add figure showing zen3geo composable DataPipes for geospatial
Illustration of zen3geo's chainable I/O readers and processors for geospatial data, designed to be ready for multi-sensor/multi-modal architectures!

Flowchart in the middle row shows STAC, vector, raster, spatial and other DataPipes making up zen3geo. Bottom row shows some of the key features of zen3geo (as of v0.6.2), and a future roadmap.

Also added logos for GeoPackage and NetCDF to the list in figures/README.md.