:art: Add figure showing demo code of a zen3geo DataPipe
Illustration of a zen3geo DataPipe from a Zarr store to Pytorch tensors. Code is shown on the left, and a graphviz plot is shown on the right, with blue arrows annotated with comments showing what each code line is doing. Also linking to full presentation slides on https://hackmd.io/@weiji14/foss4g2023oceania